Turn off esModuleInterop and fix up imports
With imports that require esModuleInterop: true, all consumers of
quill-delta must also have esModuleInterop: true. This is not always
true, specifically for @types/quill. Since @types/quill depends on
quill-delta, users will hit a Typescript compile error unless they, too,
turn on esModuleInterop.

The best solution is to turn it off and return the import statements to
the old style.