Fix retain inversion across multiple ops
Consider this base document:

```js
new Delta()
  .insert('123')
  .insert('4', { bold: true });
```

And this delta:

```js
new Delta().retain(4, { italic: true });
```

This simply applies italic to the whole document. We'd expect the
inverse to remove italic:

```js
new Delta().retain(4, { italic: null });
```

However, we currently get an inverse op with the wrong `retain` length:

```js
new Delta().retain(8, { italic: null });
```

This is because when we parse multiple ops in the inversion, we use the
original op's `retain` value (in this case `4` gets used twice), instead
of the length of the individual ops (`3` and then `1`).

This change updates the inversion to use the correct length for the
inverted `retain`.