Use ESM export statement
We used `export = Delta` until this change to keep compatible with Node.js. However, this requires `esModuleInterop` to be enabled on the user side.

This changes make it use the export statement of ESM/TypeScript, and to keep compatible with Node.js, it also replace to `module` if present.

With this change, this module should work with Node.js and TypeScript (no matter the value of `esModuleInterop`).