Remove unused lint rules
These rules that make linter looser are not used so can just remove them.
We can add them back when they become necessary.

The only usage of `any` is `AttributeMap`. This PR changes it to `unknown`,
which can represent any types like `any` but requires type assertions or control
flow based narrowing. It can be considered as a type-safe version of `any` and as a library, it's better to use it instead of `any` to avoid users from writing type-unsafe code.

However, this introduces a breaking change for TypeScript users as they have to change their code to either narrowing the types explicitly or using `as any` as we currently do for them.